added auth and crud should work now

+++ app/Http/Controllers/TreetController.php
@@ -21,6 +21,8 @@ class TreetController extends Controller
      */
     public function create(Request $request)
     {
+        $this->authorize('create', Treet::class);
+
         $treet = Treet::create([
             'app_name' => $request->app_name,
             'image_path' => $request->image_path,
@@ -34,6 +36,8 @@ class TreetController extends Controller
      */
     public function edit(Request $request, Treet $treet)
     {
+        $this->authorize('edit', Treet::class);
+
         $treet = Treet::find($request->id);
 
         $treet->update([
@@ -50,12 +54,15 @@ class TreetController extends Controller
      */
     public function destroy(Request $request, Treet $treet)
     {
+        $this->authorize('destroy', Treet::class);
+
         $treet->find($request->id)->delete();
         return $this->success($treet);
     }
     
     public function getTreetFeed(Request $request)
     {
+
         $this->validate($request, [
             'limit' => 'nullable|integer|min:6|max:90',
         ]);
